Rename components directory and update all relevant imports
